Fixes and optimisations
Fixed things that weren't working after last commit. Sending most things to global mixer (except reverb and pluck, because pluck needs a rake of instances ringing out on top of eachother). Renamed Arduino sketch folder because it cried at me. Amp levels seem to be fine coming in at around 0.5 on the master, pluck is also quiet but the reverb still seems to crackle - could be a CPU thing. Didn't bother with more Cabbage UI because the levels are fine - the two RMS meters are the left and right master channels.